+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "antd": "^4.16.13",
+    "classnames": "^2.3.1",
     "dayjs": "^1.10.6",
     "include-media": "^1.4.10",
     "qs": "^6.7.0",
@@ -19,6 +20,7 @@
   },
   "devDependencies": {
     "@craco/craco": "^6.2.0",
+    "@types/classnames": "^2.3.1",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
     "@types/qs": "^6.9.7",
