+++ src/App.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 
-import routes from "@config/routes";
 import GithubContext from "@shared/contexts/GithubContext";
 import GithubStore from "@store/GitHubStore/GitHubStore";
 import { BrowserRouter, Route, Switch, Redirect } from "react-router-dom";
+
+import routes from "./routes";
 
 const App: React.FC = () => {
   const store = new GithubStore();
@@ -13,12 +14,7 @@ const App: React.FC = () => {
       <BrowserRouter>
         <Switch>
           {routes.map((route) => (
-            <Route
-              key={route.path}
-              path={route.path}
-              exact={route.exact}
-              component={route.component}
-            />
+            <Route key={route.path} {...route} />
           ))}
           <Redirect from="*" to="/repos" />
         </Switch>
