+++ src/pages/ReposSearchPage/components/Card/Card.module.scss
@@ -7,7 +7,7 @@
   border-radius: $border-length;
   font-size: $main-font-size;
   line-height: $main-line-height;
-  border: 1px solid #D9D9D9;
+  border: 1px solid $color-grey;
   cursor: pointer;
 
   .cardContent {
@@ -33,7 +33,7 @@
         transition: color 0.5s ease-in-out;
 
         &:hover {
-          color: #208FBF;
+          color:$color-blue;
         }
       }
