+++ src/store/GitHubStore/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./GitHubStore";
+export * from "./types";
