+++ src/styles/_variables.scss
@@ -2,6 +2,7 @@ $color-yellow: #FF9432;
 $color-red: #FF5555;
 $color-grey: #7E7E7E;
 $color-white: #FFFFFF;
+$color-blue: #208FBF;
 
 $main-font-size: 16px;
 $small-font-size: 14px;
